+++ zazu/cli.py
@@ -2,7 +2,6 @@
 """Entry point for zazu."""
 import click
 import os
-import zazu.build
 import zazu.config
 import zazu.dev.commands
 import zazu.git_helper
@@ -10,8 +9,8 @@ import zazu.repo.commands
 import zazu.style
 import zazu.upgrade
 
-__author__ = "Nicholas Wiles"
-__copyright__ = "Copyright 2016"
+__author__ = 'Nicholas Wiles'
+__copyright__ = 'Copyright 2016'
 
 
 @click.group()
@@ -24,13 +23,13 @@ def cli(ctx):
 
 def init():
     """Run on startup to allow zazu to be run as a module."""
-    if __name__ == "__main__":
+    if __name__ == '__main__':
         cli()
 
 
 cli.add_command(zazu.upgrade.upgrade)
 cli.add_command(zazu.style.style)
-cli.add_command(zazu.build.build)
+cli.add_command(zazu.config.config)
 cli.add_command(zazu.dev.commands.dev)
 cli.add_command(zazu.repo.commands.repo)
 init()