+++ zazu/issue_tracker.py
@@ -1,7 +1,7 @@
 # -*- coding: utf-8 -*-
 """Issue tracker related classes."""
-__author__ = "Nicholas Wiles"
-__copyright__ = "Copyright 2016"
+__author__ = 'Nicholas Wiles'
+__copyright__ = 'Copyright 2016'
 
 
 class IssueTracker(object):
@@ -54,3 +54,11 @@ class Issue(object):
     def id(self):
         """Get the string id of the issue."""
         raise NotImplementedError('Must implement id')
+
+    def __str__(self):
+        """Return the id as the string representation."""
+        return self.id
+
+    def __repr__(self):
+        """Return the id as the string representation."""
+        return self.id