+++ zazu/plugins/docformatter_styler.py
@@ -1,13 +1,13 @@
 # -*- coding: utf-8 -*-
-"""Autopep8Styler plugin for zazu."""
+"""DocformatterStyler plugin for zazu."""
 import zazu.styler
 
 __author__ = 'Nicholas Wiles'
-__copyright__ = 'Copyright 2016'
+__copyright__ = 'Copyright 2018'
 
 
-class Autopep8Styler(zazu.styler.Styler):
-    """Autopep8 plugin for code styling."""
+class DocformatterStyler(zazu.styler.Styler):
+    """Docformatter plugin for code styling."""
 
     @staticmethod
     def default_extensions():
@@ -17,9 +17,9 @@ class Autopep8Styler(zazu.styler.Styler):
     @staticmethod
     def type():
         """Return the name of this Styler."""
-        return 'autopep8'
+        return 'docformatter'
 
     @staticmethod
     def required_options():
-        """Options required to make autopep8 take input from stdin."""
+        """Options required to make docformatter use stdin."""
         return ['-']