+++ zazu/plugins/jira_issue_tracker.py
@@ -8,8 +8,8 @@ zazu.util.lazy_import(locals(), [
     're'
 ])
 
-__author__ = "Nicholas Wiles"
-__copyright__ = "Copyright 2016"
+__author__ = 'Nicholas Wiles'
+__copyright__ = 'Copyright 2016'
 
 ZAZU_IMAGE_URL = 'http://vignette1.wikia.nocookie.net/disney/images/c/ca/Zazu01cf.png'
 ZAZU_REPO_URL = 'https://github.com/stopthatcow/zazu'
@@ -31,6 +31,7 @@ class JiraIssueTracker(zazu.issue_tracker.IssueTracker):
         self._default_project = default_project
         self._components = components
         self._jira_handle = None
+        self._user = None
 
     def connect(self):
         """Get handle to ensure that JIRA credentials are in place."""
@@ -48,6 +49,12 @@ class JiraIssueTracker(zazu.issue_tracker.IssueTracker):
         """Get the url to open to display the issue."""
         normalized_id = self.validate_id_format(id)
         return '{}/browse/{}'.format(self._base_url, normalized_id)
+
+    def user(self):
+        """Get username of authenticated user."""
+        if self._user is None:
+            self._user = self._jira().current_user()
+        return self._user
 
     def issue(self, id):
         """Get an issue by id."""
@@ -92,6 +99,10 @@ class JiraIssueTracker(zazu.issue_tracker.IssueTracker):
         issues = self._jira().search_issues('assignee={}'.format(self._username))
         return [JiraIssueAdaptor(i) for i in issues]
 
+    def assign_issue(self, issue, user):
+        """Assign an issue to a user."""
+        self._jira().assign_issue(issue._jira_issue, user)
+
     def default_project(self):
         """JIRA project associated with this tracker."""
         return self._default_project
@@ -115,6 +126,7 @@ class JiraIssueTracker(zazu.issue_tracker.IssueTracker):
 
         Returns:
             normalized id string
+
         """
         components = id.split('-', 1)
         number = components.pop()
@@ -145,7 +157,7 @@ class JiraIssueTracker(zazu.issue_tracker.IssueTracker):
     @staticmethod
     def type():
         """Return the name of this IssueTracker type."""
-        return 'Jira'
+        return 'jira'
 
 
 class JiraIssueAdaptor(zazu.issue_tracker.Issue):
@@ -200,7 +212,3 @@ class JiraIssueAdaptor(zazu.issue_tracker.Issue):
     def id(self):
         """Get the string id of the issue."""
         return self._jira_issue.key
-
-    def __str__(self):
-        """Return the id as the string representation."""
-        return self.id